requirements.txt
alembic==1.4.3
basilica==0.2.8
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
Pillow==8.0.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.20
threadpoolctl==2.1.0
tweepy==3.9.0
urllib3==1.25.11
Werkzeug==1.0.1
